koordlet: fix build error on macOS caused by GPU
